+++ src/Products/PloneMeeting/vocabularies.py
@@ -1015,30 +1015,33 @@ class ItemOptionalAdvicesVocabulary(object):
                                              context=request)
             return value_to_display
 
-        def _insert_term_and_users(res, term_value, term_title):
+        def _insert_term_and_users(res, term_value, term_title, add_users=True):
             """ """
             term = SimpleTerm(term_value, term_value, term_title)
             term.sortable_title = term_title
             res.append(term)
             org_uid = term_value.split('__rowid__')[0]
-            if org_uid in selectableAdviserUsers:
-                advisers_group = get_plone_group(org_uid, "advisers")
-                for user_id in advisers_group.getGroupMemberIds():
+            if add_users:
+                user_ids = []
+                if org_uid in selectableAdviserUsers:
+                    user_ids += get_plone_group(org_uid, "advisers").getGroupMemberIds()
+                # manage missing user ids here so term is grouped with the org term
+                prefix = term_value + '__userid__'
+                missing_user_ids = [oa.replace(prefix, '') for oa in context.getOptionalAdvisers()
+                                    if oa.startswith(prefix)]
+                user_ids += missing_user_ids
+                # manage users in a separate list so we sort it before appending to global res
+                res_users = []
+                for user_id in user_ids:
                     user_term_value = "{0}__userid__{1}".format(term_value, user_id)
                     user_title = safe_unicode(tool.getUserName(user_id))
                     user_term = SimpleTerm(user_term_value, user_term_value, user_title)
                     user_term.sortable_title = u"{0} ({1})".format(term_title, user_title)
-                    res.append(user_term)
+                    res_users.append(user_term)
+                res_users = humansorted(res_users, key=attrgetter('title'))
+                res += res_users
             return
 
-        def _getNonDelayAwareAdvisers_cachekey(method, cfg):
-            '''cachekey method for self._getNonDelayAwareAdvisers.'''
-            # this volatile is invalidated when plonegroup config changed
-            date = get_cachekey_volatile(
-                '_users_groups_value')
-            return date, repr(cfg), cfg.modified()
-
-        @ram.cache(_getNonDelayAwareAdvisers_cachekey)
         def _getNonDelayAwareAdvisers(cfg):
             """Separated so it can be cached."""
             resNonDelayAwareAdvisers = []
@@ -1087,23 +1090,42 @@ class ItemOptionalAdvicesVocabulary(object):
                                           [org_infos.token for org_infos in resDelayAwareAdvisers]
                 for optionalAdviser in optionalAdvisers:
                     if optionalAdviser not in optionalAdvisersInVocab:
+                        user_id = org = None
+                        org_uid = optionalAdviser
+                        if '__userid__' in optionalAdviser:
+                            org_uid, user_id = optionalAdviser.split('__userid__')
                         if '__rowid__' in optionalAdviser:
-                            org_uid, row_id = decodeDelayAwareId(optionalAdviser)
+                            org_uid, row_id = decodeDelayAwareId(org_uid)
                             delay = cfg._dataForCustomAdviserRowId(row_id)['delay']
                             delay_label = context.adviceIndex[org_uid]['delay_label']
                             org = get_organization(org_uid)
                             if not org:
                                 continue
-                            org_title = org.get_full_title()
-                            value_to_display = _displayDelayAwareValue(delay_label, org_title, delay)
-                            _insert_term_and_users(
-                                resDelayAwareAdvisers, optionalAdviser, value_to_display)
+                            value_to_display = _displayDelayAwareValue(
+                                delay_label, org.get_full_title(), delay)
+                            if not user_id:
+                                _insert_term_and_users(
+                                    resDelayAwareAdvisers,
+                                    optionalAdviser,
+                                    value_to_display,
+                                    add_users=False)
                         else:
-                            org = get_organization(optionalAdviser)
+                            org = get_organization(org_uid)
                             if not org:
                                 continue
-                            _insert_term_and_users(
-                                resNonDelayAwareAdvisers, optionalAdviser, org.get_full_title())
+                            if not user_id:
+                                _insert_term_and_users(
+                                    resNonDelayAwareAdvisers,
+                                    optionalAdviser,
+                                    org.get_full_title(),
+                                    add_users=False)
+                        # it is a userid, add a special value including the org title
+                        if org and user_id:
+                            user_term_title = safe_unicode(
+                                "{0} ({1})".format(org.get_full_title(), tool.getUserName(user_id)))
+                            user_term = SimpleTerm(optionalAdviser, optionalAdviser, user_term_title)
+                            user_term.sortable_title = user_term_title
+                            resDelayAwareAdvisers.append(user_term)
 
         # now create the listing
         # sort elements by value before potentially prepending a special value here under